Tools/discord_agent_bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
watchdog>=3.0.0
PyYAML>=6.0
